requirements.txt
pytesseract
pillow
pyautogui
keyboard
tkinter
easyocr
